refactor (BACK-END): adequando aos testes unitários

+++ micro-services/cliente-service/src/main/java/br/com/tads/dac/clienteservice/model/dto/ClienteTransacoesResponseDTO.java
@@ -16,7 +16,7 @@ public class ClienteTransacoesResponseDTO {
 
     private String codigo;
 
-    private Integer saldoMilhas;
+    private Long saldoMilhas;
 
     private List<TransacaoMilhasDTO> transacoes;
 }
